refactor: move question component stuff to for-students-FAQ.js
