+++ src/for-students-FAQ/for-students-FAQ.js
@@ -1,43 +1,22 @@
 import { Accordion } from "@mantine/core";
-import "./question.css";
+import "./for-students-FAQ.css";
 
-const groceries = [
-  {
-    emoji: "🍎",
-    value: "Apples",
-    description:
-      "Crisp and refreshing fruit. Apples are known for their versatility and nutritional benefits. They come in a variety of flavors and are great for snacking, baking, or adding to salads.",
-  },
-  {
-    emoji: "🍌",
-    value: "Bananas",
-    description:
-      "Naturally sweet and potassium-rich fruit. Bananas are a popular choice for their energy-boosting properties and can be enjoyed as a quick snack, added to smoothies, or used in baking.",
-  },
-  {
-    emoji: "🥦",
-    value: "Broccoli",
-    description:
-      "Nutrient-packed green vegetable. Broccoli is packed with vitamins, minerals, and fiber. It has a distinct flavor and can be enjoyed steamed, roasted, or added to stir-fries.",
-  },
-];
-
-function Demo() {
-  const items = groceries.map((item) => (
-    <Accordion.Item key={item.value} value={item.value}>
+function FaqCard({ faq_info }) {
+  const items = faq_info.map((faq) => (
+    <Accordion.Item key={faq.value} value={faq.value}>
       <Accordion.Control className="question">
-        <b>{item.value}</b>
+        <b>{faq.value}</b>
       </Accordion.Control>
 
-      <Accordion.Panel>{item.description}</Accordion.Panel>
+      <Accordion.Panel>{faq.description}</Accordion.Panel>
     </Accordion.Item>
-  ))
+  ));
 
   return (
     <Accordion multiple variant="filled" defaultValue="Apples">
       {items}
     </Accordion>
-  )
+  );
 }
 
-export { Demo };
+export default FaqCard;
